+++ keyfence/addon.py
@@ -232,7 +232,7 @@ class KeyFence:
             return
 
         if self.config.notice:
-            new_text = add_notice(new_text, host)
+            new_text = add_notice(new_text, host, self.config.mode)
         flow.request.set_text(new_text)
         if mapping:
             flow.metadata[MAPPING_KEY] = mapping